test(e2e): moved utils to src folder from test folder
