+++ packages/e2e/src/util/index.ts
@@ -1,2 +1,3 @@
 export * from './StubKeyAgent';
 export * from './localNetworkChainId';
+export * from './util';
