+++ packages/e2e/test/artillery/wallet-restoration/WalletRestoration.ts
@@ -1,12 +1,20 @@
+import { AddressType, GroupedAddress, util } from '@cardano-sdk/key-management';
 import { AddressesModel, WalletVars } from './types';
 import { Cardano } from '@cardano-sdk/core';
 import { FunctionHook } from '../artillery';
 import { Pool } from 'pg';
-import { StubKeyAgent, getEnv, getWallet, walletVariables } from '../../../src';
+import { StubKeyAgent, getEnv, getWallet, waitForWalletStateSettle, walletVariables } from '../../../src';
 import { findAddressesWithRegisteredStakeKey } from './queries';
 import { logger } from '@cardano-sdk/util-dev';
-import { mapToGroupedAddress, waitForWalletStateSettle } from '../../util';
-import { util } from '@cardano-sdk/key-management';
+
+export const mapToGroupedAddress = (addrModel: AddressesModel): GroupedAddress => ({
+  accountIndex: 0,
+  address: Cardano.PaymentAddress(addrModel.address),
+  index: 0,
+  networkId: addrModel.address.startsWith('addr_test') ? Cardano.NetworkId.Testnet : Cardano.NetworkId.Mainnet,
+  rewardAccount: Cardano.RewardAccount(addrModel.stake_address),
+  type: AddressType.External
+});
 
 const env = getEnv([
   ...walletVariables,
