+++ packages/e2e/test/wallet/PersonalWallet/mint.test.ts
@@ -2,10 +2,17 @@ import { Cardano, nativeScriptPolicyId } from '@cardano-sdk/core';
 import { FinalizeTxProps, PersonalWallet } from '@cardano-sdk/wallet';
 import { InitializeTxProps } from '@cardano-sdk/tx-construction';
 import { KeyRole, util } from '@cardano-sdk/key-management';
-import { burnTokens, createStandaloneKeyAgent, submitAndConfirm, walletReady } from '../../util';
+import {
+  burnTokens,
+  createStandaloneKeyAgent,
+  getEnv,
+  getWallet,
+  submitAndConfirm,
+  walletReady,
+  walletVariables
+} from '../../../src';
 import { createLogger } from '@cardano-sdk/util-dev';
 import { filter, firstValueFrom, map, take } from 'rxjs';
-import { getEnv, getWallet, walletVariables } from '../../../src';
 import { isNotNil } from '@cardano-sdk/util';
 
 const env = getEnv(walletVariables);
