+++ packages/e2e/test/wallet/PersonalWallet/txChainHistory.test.ts
@@ -1,9 +1,8 @@
 import { PersonalWallet } from '@cardano-sdk/wallet';
 import { filter, firstValueFrom, map, take } from 'rxjs';
-import { getEnv, getWallet, walletVariables } from '../../../src';
+import { getEnv, getWallet, normalizeTxBody, walletReady, walletVariables } from '../../../src';
 import { isNotNil } from '@cardano-sdk/util';
 import { logger } from '@cardano-sdk/util-dev';
-import { normalizeTxBody, walletReady } from '../../util';
 
 const env = getEnv(walletVariables);
